Destroy organization users and databases

+++ app/models/user.rb
@@ -189,15 +189,6 @@ class User < Sequel::Model
         end
       end
 
-      # Right now, cannot delete users with entities shared with other users or the org.
-      can_delete = true
-      CartoDB::Permission.where(owner_id: self.id).each { |permission|
-        can_delete = can_delete && permission.acl.empty?
-      }
-      unless can_delete
-        raise CartoDB::BaseCartoDBError.new('Cannot delete user, has shared entities')
-      end
-
       has_organization = true
     end
 
@@ -228,6 +219,10 @@ class User < Sequel::Model
     # Invalidate user cache
     invalidate_varnish_cache
 
+    CartoDB::Permission.where(owner_id: self.id).each { |permission|
+      permission.destroy
+    }
+
     # Delete the DB or the schema
     if has_organization
       drop_organization_user(org_id, is_owner = !@org_id_for_org_wipe.nil?) unless error_happened
@@ -235,7 +230,12 @@ class User < Sequel::Model
       if User.where(:database_name => self.database_name).count > 1
         raise CartoDB::BaseCartoDBError.new('The user is not supposed to be in a organization but another user has the same database_name. Not dropping it')
       else
-        drop_database_and_user unless error_happened
+        if !error_happened
+          Thread.new do
+            drop_database_and_user
+          end.join
+          monitor_user_notification
+        end
       end
     end
 
@@ -262,7 +262,7 @@ class User < Sequel::Model
     end
   end
 
-  # Org users share the same db, so must only delete the schema
+  # Org users share the same db, so must only delete the schema unless he's the owner
   def drop_organization_user(org_id, is_owner=false)
     raise CartoDB::BaseCartoDBError.new('Tried to delete an organization user without org id') if org_id.nil?
 
@@ -299,21 +299,23 @@ class User < Sequel::Model
       conn.run("REVOKE ALL ON DATABASE \"#{database_name}\" FROM \"#{database_username}\"")
       conn.run("REVOKE ALL ON DATABASE \"#{database_name}\" FROM \"#{database_public_username}\"")
       conn.run("DROP USER \"#{database_public_username}\"")
-      conn.run("DROP USER \"#{database_username}\"")
+      if is_owner
+        drop_database_and_user if is_owner
+      else
+        conn.run("DROP USER \"#{database_username}\"")
+      end
+
     end.join
 
     monitor_user_notification
   end
 
   def drop_database_and_user
-    Thread.new do
-      conn = self.in_database(as: :cluster_admin)
-      conn.run("UPDATE pg_database SET datallowconn = 'false' WHERE datname = '#{database_name}'")
-      User.terminate_database_connections(database_name, database_host)
-      conn.run("DROP DATABASE \"#{database_name}\"")
-      conn.run("DROP USER \"#{database_username}\"")
-    end.join
-    monitor_user_notification
+    conn = self.in_database(as: :cluster_admin)
+    conn.run("UPDATE pg_database SET datallowconn = 'false' WHERE datname = '#{database_name}'")
+    User.terminate_database_connections(database_name, database_host)
+    conn.run("DROP DATABASE \"#{database_name}\"")
+    conn.run("DROP USER \"#{database_username}\"")
   end
 
   def self.terminate_database_connections(database_name, database_host)
